# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LMS-AI API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/courses/{course_id}/analytics": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Course Analytics",
        "operationId": "course_analytics_api_v1_courses__course_id__analytics_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseAnalytics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/courses/{course_id}/tutor-feedback": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Course Tutor Feedback",
        "description": "Câu trả lời AI Tutor bị học viên chê trong khóa (không kèm tên học viên).",
        "operationId": "course_tutor_feedback_api_v1_courses__course_id__tutor_feedback_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Course Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutorFeedbackPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CourseAnalytics": {
        "properties": {
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "enrollments": {
            "type": "integer",
            "title": "Enrollments"
          },
          "completed_enrollments": {
            "type": "integer",
            "title": "Completed Enrollments"
          },
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/LessonStat"
            },
            "type": "array",
            "title": "Lessons"
          },
          "quizzes": {
            "items": {
              "$ref": "#/components/schemas/QuizStat"
            },
            "type": "array",
            "title": "Quizzes"
          },
          "tutor": {
            "$ref": "#/components/schemas/TutorStat"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "enrollments",
          "completed_enrollments",
          "lessons",
          "quizzes",
          "tutor"
        ],
        "title": "CourseAnalytics"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LessonStat": {
        "properties": {
          "lesson_id": {
            "type": "string",
            "format": "uuid",
            "title": "Lesson Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "section_title": {
            "type": "string",
            "title": "Section Title"
          },
          "done_count": {
            "type": "integer",
            "title": "Done Count"
          },
          "completion_rate": {
            "type": "number",
            "title": "Completion Rate"
          }
        },
        "type": "object",
        "required": [
          "lesson_id",
          "title",
          "section_title",
          "done_count",
          "completion_rate"
        ],
        "title": "LessonStat"
      },
...
      "QuizStat": {
        "properties": {
          "quiz_id": {
            "type": "string",
            "format": "uuid",
            "title": "Quiz Id"
          },
          "lesson_id": {
            "type": "string",
            "format": "uuid",
            "title": "Lesson Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/QuizStatus"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "students": {
            "type": "integer",
            "title": "Students"
          },
          "avg_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Score"
          },
          "pass_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pass Rate"
          }
        },
        "type": "object",
        "required": [
          "quiz_id",
          "lesson_id",
          "title",
          "status",
          "attempts",
          "students",
          "avg_score",
          "pass_rate"
        ],
        "title": "QuizStat"
      },
      "QuizStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published"
        ],
        "title": "QuizStatus"
      },
...
      "TutorFeedbackItem": {
        "properties": {
          "message_id": {
            "type": "string",
            "format": "uuid",
            "title": "Message Id"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "refused": {
            "type": "boolean",
            "title": "Refused"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "course_title": {
            "type": "string",
            "title": "Course Title"
          },
          "course_slug": {
            "type": "string",
            "title": "Course Slug"
          },
          "lesson_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lesson Id"
          },
          "lesson_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lesson Title"
          },
          "student_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Name"
          }
        },
        "type": "object",
        "required": [
          "message_id",
          "question",
          "answer",
          "refused",
          "created_at",
          "course_id",
          "course_title",
          "course_slug",
          "lesson_id",
          "lesson_title",
          "student_name"
        ],
        "title": "TutorFeedbackItem",
        "description": "Một câu trả lời AI Tutor bị học viên bấm 👎 (D1), kèm câu hỏi ngay trước nó."
      },
      "TutorFeedbackPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TutorFeedbackItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "title": "TutorFeedbackPage"
      },
      "TutorStat": {
        "properties": {
          "sessions": {
            "type": "integer",
            "title": "Sessions"
          },
          "questions": {
            "type": "integer",
            "title": "Questions"
          },
          "refused_answers": {
            "type": "integer",
            "title": "Refused Answers"
          }
        },
        "type": "object",
        "required": [
          "sessions",
          "questions",
          "refused_answers"
        ],
        "title": "TutorStat"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}